+++ src/selectors/questionnaire/answer.ts
@@ -1,13 +1,5 @@
 import * as model from '../../stores/questionnaire';
 
-export interface Question {
-    readonly id: model.Id;
-    readonly number: number;
-    readonly text: string;
-    readonly explanation?: string;
-    readonly answers: ReadonlyArray<Answer>;
-}
-
 export interface Answer {
     readonly id: model.Id;
     readonly text: string;
